Add support for extends as array of strings.
TypeScript 5.0 added support for defining "extends" as an array of
strings. This commit adds support for this use case.

It's important to note that even with this change, "baseUrl" and
"paths" are still always being completely overwritten if a later
tsconfig redefines any of those values.
This might be confusing because a tsconfig may define
"baseUrl=value1" and its own set of "paths" based on that baseUrl,
but if a later tsconfig defines its own "baseUrl=value2", the
overall config ends up becoming "baseUrl=value2" with the "paths"
from the first config.

This behaviour hasn't changed even when "extends" is an array of
strings, so this commit maintains this behaviour.